+++ backends/arm/_passes/BUCK
@@ -74,6 +74,7 @@ fbcode_target(
         "//executorch/backends/arm/tosa/dialect:lib",
         "//executorch/backends/transforms:canonicalize_view_copy_permute_pass",
         "//executorch/backends/transforms:fuse_duplicate_users_pass",
+        "//executorch/backends/transforms:propagate_view_copy_permute_pass",
         "//executorch/backends/transforms:fuse_identical_input_transforms_pass",
         "//executorch/backends/transforms:convert_conv1d_to_conv2d_pass",
         "//executorch/backends/transforms:fuse_view_copy",
